Accept return codes 0 and 1 for Debian grep call
In the command used to check for updates on Debian-based systems, a
`grep -c` call is used. In case that doesn't match, grep will print "0"
on stdout (which is desired and expected), but will also return with an
error code of 1 instead of 0. Expect this by explicitly allowing error
codes 0 and 1 and avoid the exception otherwise thrown by sh.

This fixes issue #1.

Signed-off-by: Kai Blin <kai@samba.org>